monitor/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
aiocron==1.4
    # via -r requirements/common.in
aiohttp==3.7.4.post0
    # via
    #   -r requirements/common.in
    #   aioresponses
aiokafka==0.7.0
    # via -r requirements/common.in
aioresponses==0.7.2
    # via -r requirements/dev.in
appdirs==1.4.4
    # via black
appnope==0.1.2
    # via ipython
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   pytest
backcall==0.2.0
    # via ipython
black==21.5b0
    # via -r requirements/dev.in
chardet==4.0.0
    # via aiohttp
click==7.1.2
    # via black
croniter==1.0.13
    # via aiocron
decorator==5.0.7
    # via ipython
flake8==3.9.2
    # via -r requirements/dev.in
idna==3.1
    # via yarl
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via traitlets
ipython==7.23.1
    # via -r requirements/dev.in
isort==5.8.0
    # via -r requirements/dev.in
jedi==0.18.0
    # via ipython
kafka-python==2.0.2
    # via aiokafka
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via flake8
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via -r requirements/dev.in
packaging==20.9
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pydantic==1.8.1
    # via -r requirements/common.in
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest-asyncio==0.15.1
    # via -r requirements/dev.in
pytest==6.2.4
    # via
    #   -r requirements/dev.in
    #   pytest-asyncio
python-dateutil==2.8.1
    # via croniter
pytz==2021.1
    # via tzlocal
pyyaml==5.4.1
    # via -r requirements/common.in
regex==2021.4.4
    # via black
six==1.16.0
    # via python-dateutil
toml==0.10.2
    # via
    #   black
    #   pytest
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typed-ast==1.4.3
    # via mypy
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   mypy
    #   pydantic
tzlocal==2.1
    # via aiocron
wcwidth==0.2.5
    # via prompt-toolkit
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
